requirements.txt
fastapi
influxdb-client
uvicorn
python-dotenv